+++ cs/Markdown/Markdown.csproj
@@ -60,6 +60,7 @@
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Tag.cs" />
+    <Compile Include="TagsMarker.cs" />
     <Compile Include="TagWithToken.cs" />
     <Compile Include="Token.cs" />
   </ItemGroup>
