+++ cs/Markdown/Md.cs
@@ -22,7 +22,7 @@ namespace Markdown
         public string Render(string text)
         {
             var tokens = Parser.ParseTextToTokens(text, Tags);
-            var tagWithTokens = TokensToTagWithTokens(tokens, text);
+            var tagWithTokens = TagsMarker.TokensToTagWithTokens(tokens, text, Tags);
             var htmlText = ConvertTokensToHtml(tagWithTokens, text);
             return htmlText;
         }
@@ -47,67 +47,6 @@ namespace Markdown
                 }
             }
             return stringBuilder.ToString();
-        }
-
-        private List<TagWithToken> TokensToTagWithTokens(List<Token> tokens, string text)
-        {
-            var openingTagList = new List<TagWithToken>();
-            var tagWithTokens = new List<TagWithToken>();
-            foreach (var token in tokens)
-            {
-                if (token.IsTag)
-                {
-                    var tag = Tags.Find(t => t.MatchTagAndTokenCompletely(token.Index, token.Length, text));
-                    var tagWithToken = new TagWithToken(tag, token);
-                    if (tagWithToken.CanBeTag(text))
-                    {
-                        tagWithTokens.Add(tagWithToken);
-                        HandlerTag(openingTagList, tagWithTokens, tagWithToken, text);
-                    }
-                    else
-                        tagWithTokens.Add(new TagWithToken(null, token));
-                }
-                else
-                {
-                    tagWithTokens.Add(new TagWithToken(null, token));
-                }
-            }
-            return tagWithTokens;
-        }
-
-        private void HandlerTag(List<TagWithToken> openingTagList, List<TagWithToken> tagWithTokens, TagWithToken tagWithToken, string text)
-        {
-            if (tagWithToken.CanTagBeClosing(text))
-            {
-                var indexOpeningTag = openingTagList.FindLastIndex(tag => tagWithToken.Tag.MarkdownTag == tag.Tag.MarkdownTag && tag.CanTagBeOpening(text));
-                if (indexOpeningTag >= 0)
-                {
-                    openingTagList[indexOpeningTag].IsOpen = true;
-                    tagWithToken.IsClose = true;
-                    openingTagList.RemoveRange(indexOpeningTag, openingTagList.Count - indexOpeningTag);
-
-                    switch (tagWithToken.Tag.MarkdownTag)
-                    {
-                        case "_":
-                            var i = tagWithTokens.Count - 1;
-                            while (tagWithTokens[i] != tagWithTokens[indexOpeningTag])
-                            {
-                                if (tagWithTokens[i].IsTag && tagWithTokens[i].Tag.MarkdownTag == "__")
-                                {
-                                    tagWithTokens[i].IsClose = false;
-                                    tagWithTokens[i].IsOpen = false;
-                                }
-                                i--;
-                            }
-                            break;
-                    }
-                    return;
-                }
-            }
-            if (tagWithToken.CanTagBeOpening(text))
-            {
-                openingTagList.Add(tagWithToken);
-            }
         }
     }
 }
